requirements.txt
numpy
matplotlib
opencv-python
transformers
pillow
requests
datasets
torch
torchvision
timm


